Added Documentation to the Project
Added docs as .yml in the staging and prod models. Additionally, I spruced up the README based on the jaffle_shop project.

I added tests to the schema in the documentation, and fixed two emails both with two occurrences each. James as a first name, and Gonzalez/Fowler as last names. Schema tests ftw!